Split repeat commands into separate files.
