+++ sway/meson.build
@@ -90,7 +90,8 @@ sway_sources = files(
 	'commands/input/middle_emulation.c',
 	'commands/input/natural_scroll.c',
 	'commands/input/pointer_accel.c',
-	'commands/input/repeat.c',
+	'commands/input/repeat_delay.c',
+	'commands/input/repeat_rate.c',
 	'commands/input/scroll_method.c',
 	'commands/input/tap.c',
 	'commands/input/xkb_layout.c',
